requirements_txt.txt
# AI Workflow Builder Dependencies

# Core AI and API
openai==0.28.1
anthropic==0.3.11

# Data processing
pandas==2.0.3
numpy==1.24.3

# Async and HTTP
aiohttp==3.8.5
asyncio-throttle==1.0.2

# File handling
openpyxl==3.1.2
python-docx==0.8.11

# Configuration and environment
python-dotenv==1.0.0
pydantic==2.3.0

# CLI and UI
click==8.1.7
rich==13.5.2
streamlit==1.26.0

# Logging and monitoring
loguru==0.7.0

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1

# Development tools
black==23.7.0
flake8==6.0.0
mypy==1.5.1